+++ src/Persistence/tests/test_mapping.py
@@ -19,3 +19,17 @@ class PersistentMappingTests(unittest.TestCase):
         m = self._makeOne()
         m.__setstate__({'_container': {'x': 1, 'y': 2}})
         self.assertEqual(sorted(m.items()), [('x', 1), ('y', 2)])
+
+    def test_subclass(self):
+        klass = self._getTargetClass()
+
+        class Dummy(klass):
+            def __init__(self, *args, **kw):
+                self._value = []
+                klass.__init__(self, *args, **kw)
+
+        self.assertTrue(issubclass(Dummy, klass))
+
+        dummy = Dummy()
+        self.assertTrue(isinstance(dummy, Dummy))
+        self.assertTrue(isinstance(dummy, klass))